pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "lfind"
version = "0.2.0"
description = "A natural language file finder using LLMs"
authors = [
    {name = "Samuel Buban", email = "samuelbuban@gmail.com"}
]
license = {file = "LICENSE"}
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "openai>=1.61.1",
    "python-dotenv>=1.0.1",
    "faiss-cpu>=1.7.0",
    "sentence-transformers>=2.2.0",
    "PyPDF2>=3.0.0",  # For PDF text extraction
]

[project.urls]
Homepage = "https://github.com/Mahrkeenerh/lfind"
Repository = "https://github.com/Mahrkeenerh/lfind.git"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
include = ["lfind*"]

[project.scripts]
lfind = "lfind.main:main"